:loud_sound: 添加日志

+++ esystem/src/main/java/io/kid1999/esystem/dao/CollectionDao.java
@@ -18,5 +18,5 @@ public interface CollectionDao extends BaseMapper<Collection> {
 
     @Select("select goods.id, goods.user_id, goods_name, img_url, goods_condition, description, want, exchange_addr, remarks, goods_status, number_of_clicked, goods.create_date, goods.deleted, price " +
             "from goods left join collection on collection.user_id = goods.user_id where collection.user_id = #{userId} and collection.deleted = 0 and goods.deleted = 0")
-    List<Goods> selectCollectionsByUserId(int userId);
+    List<Goods> selectCollectionsByUserId(long userId);
 }
